Create Chapter 1-Exercise:1
Exercise 1-1. Load the image dog.jpg and compute the full SVD. Choose a rank r < m and confirm that the matrix U∗U is the r × r identity matrix. Now confirm that UU∗ is not the identity matrix. Compute the norm of the error between UU∗ and the n × n identity matrix as the rank r varies from 1 to n and plot the
error